% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iEN.R
\docType{class}
\name{iEN-class}
\alias{iEN-class}
\alias{iEN}
\title{An S4 class representative of the immunological Elastic Net model}
\description{
An S4 class representative of the immunological Elastic Net model
}
\section{Slots}{

\describe{
\item{\code{cv.preds}}{out-of-sample predictions made during k-Fold cross validation.}

\item{\code{cv.eval}}{out-of-sample evaluation of predictive analysis, as defined by cv.kEN parameter eval, this estimates the models ability at predicting unseen data.}

\item{\code{betas}}{matrix of model coefficients for each optimized out-of-sample model of the K-fold cross validation.}

\item{\code{dfs}}{vector of values corresponding to the number of non-zero elements in each out-of-sample model}

\item{\code{nobs}}{number of observations the provided dataset contained.}

\item{\code{nfeat}}{number of features the provided dataset contained.}

\item{\code{opt.params}}{optimal parameters (alpha, phi, lambda) for each out-of-sample model generated from the initial search parameters provided.}

\item{\code{call}}{function call as string from the cv.kEN call which generated this particular object.}
}}

