% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_iEN.R
\docType{methods}
\name{predict,iEN-method}
\alias{predict,iEN-method}
\title{predicts with a iEN object, defualt of the model is to predict the new data with the mean coefficient model over each out of sample instance. With the default phi applied to to newx being the mean phi from optimal models found during CV.}
\usage{
\S4method{predict}{iEN}(object, newx, priors, phi)
}
\arguments{
\item{object}{object of type iEN}

\item{newx}{new data to predict}

\item{priors}{priors indicating how phi will be applied, should be identical to the priors used during cross-validation.}

\item{phi}{The amount of prioritization to use on newx, this should be the same phi as is used to construct the iEN model during prediction. Default for this method is to apply the mean phi from out of sample models.}
}
\value{
vector of predicted values using object to predict newx
}
\description{
predicts with a iEN object, defualt of the model is to predict the new data with the mean coefficient model over each out of sample instance. With the default phi applied to to newx being the mean phi from optimal models found during CV.
}
\examples{
data(test_data)

alphaGrid <- seq(0,1, length.out=2)
phiGrid <- exp(seq(log(1),log(10), length.out=2))
nlambda <- 3
ncores <- 2
eval <- "RSS" 
family <- "gaussian"
intercept <- TRUE
standardize <- TRUE
center <- TRUE

model <- cv_iEN(X, Y, foldid, alphaGrid, phiGrid, nlambda, NULL, priors, ncores, eval, family, intercept, standardize, center)
predict(model, X, priors)
}
