% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_iEN.R
\docType{methods}
\name{print,iEN-method}
\alias{print,iEN-method}
\title{Prints iEN object in a formatted way using the knitr package, displays optimized hyper parameters for each out of sample instance during cross-validated gridsearch.}
\usage{
\S4method{print}{iEN}(x, ...)
}
\arguments{
\item{x}{object of type iEN}

\item{...}{additional parameters fed to print}
}
\value{
formatted print
}
\description{
Prints iEN object in a formatted way using the knitr package, displays optimized hyper parameters for each out of sample instance during cross-validated gridsearch.
}
\examples{
data(test_data)

alphaGrid <- seq(0,1, length.out=2)
phiGrid <- exp(seq(log(1),log(10), length.out=2))
nlambda <- 3
ncores <- 2
eval <- "RSS" 
family <- "gaussian"
intercept <- TRUE
standardize <- TRUE
center <- TRUE

model <- cv_iEN(X, Y, foldid, alphaGrid, phiGrid, nlambda, NULL, priors, ncores, eval, family, intercept, standardize, center)
print(model)
}
